Major update to sub1
Updated sub1 with motor control and moved folder
