+++ SUB1/README.md
@@ -1,0 +1,2 @@
+# _LETREP2 Subsystem 1_
+This is the subsystem 1 for LETREP2's Rehablititation device. This subsystem controls inputs to a clearpath motor to allow for actuation and calibration of the motor.
