+++ SUB1/main/CMakeLists.txt
@@ -1,0 +1,2 @@
+idf_component_register(SRCS "hlfb.c" "message_handling.c" "main.c"
+                    INCLUDE_DIRS ".")
