+++ SUB1/main/component.mk
