--- SUB1/sample_project/main/CMakeLists.txt
@@ -1,2 +1,0 @@
-idf_component_register(SRCS "main.c"
-                    INCLUDE_DIRS ".")
